fix: Deploy lambdas using pre-built packages
The existing method has Terraform building the package during it's run,
requiring the runner to have specific Python versions installed. Without
these versions installed the build fails, and the Terraform Cloud agents
do not have current Python versions baked in.

To circumvent this, we update the module to skip the build process and
expect a pre-built zip, which we will build outside of Terraform.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>